Struct romio::net::tcp::TcpStream

pub struct TcpStream { /* fields omitted */ }

An I/O object representing a TCP stream connected to a remote endpoint.

A TCP stream can either be created by connecting to an endpoint, via the connect method, or by accepting a connection from a listener.

Methods

impl TcpStream

pub fn connect(addr: &SocketAddr) -> ConnectFuture

Create a new TCP stream connected to the specified address.

This function will create a new TCP socket and attempt to connect it to the addr provided. The returned future will be resolved once the stream has successfully connected, or it will return an error if one occurs.

pub fn from_std(stream: TcpStream, handle: &Handle) -> Result<TcpStream>

Create a new TcpStream from a net::TcpStream.

This function will convert a TCP stream created by the standard library to a TCP stream ready to be used with the provided event loop handle. Use Handle::default() to lazily bind to an event loop, just like connect does.

pub fn connect_std(
stream: TcpStream,
addr: &SocketAddr,
handle: &Handle
) -> ConnectFuture

Creates a new TcpStream from the pending socket inside the given std::net::TcpStream, connecting it to the address specified.

This constructor allows configuring the socket before it's actually connected, and this function will transfer ownership to the returned TcpStream if successful. An unconnected TcpStream can be created with the net2::TcpBuilder type (and also configured via that route).

The platform specific behavior of this function looks like:

On Unix, the socket is placed into nonblocking mode and then a connect call is issued.

On Windows, the address is stored internally and the connect operation is issued when the returned TcpStream is registered with an event loop. Note that on Windows you must bind a socket before it can be connected, so if a custom TcpBuilder is used it should be bound (perhaps to INADDR_ANY) before this method is called.

pub fn poll_read_ready(&self, lw: &LocalWaker) -> Poll<Result<Ready>>

Check the TCP stream's read readiness state.

The mask argument allows specifying what readiness to notify on. This can be any value, including platform specific readiness, except writable. HUP is always implicitly included on platforms that support it.

If the resource is not ready for a read then Poll::Pending is returned and the current task is notified once a new event is received.

The stream will remain in a read-ready state until calls to poll_read return NotReady.

Panics

This function panics if:

ready includes writable.

called from outside of a task context.

pub fn poll_write_ready(&self, lw: &LocalWaker) -> Poll<Result<Ready>>

Check the TCP stream's write readiness state.

This always checks for writable readiness and also checks for HUP readiness on platforms that support it.

If the resource is not ready for a write then Poll::Pending is returned and the current task is notified once a new event is received.

The I/O resource will remain in a write-ready state until calls to poll_write return NotReady.

Panics

This function panics if called from outside of a task context.

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this stream is bound to.

pub fn peer_addr(&self) -> Result<SocketAddr>

Returns the remote address that this stream is connected to.

pub fn poll_peek(
&mut self,
buf: &mut [u8],
lw: &LocalWaker
) -> Poll<Result<usize>>

Receives data on the socket from the remote address to which it is connected, without removing that data from the queue. On success, returns the number of bytes peeked.

Successive calls return the same data. This is accomplished by passing MSG_PEEK as a flag to the underlying recv system call.

Return

On success, returns Ok(Async::Ready(num_bytes_read)).

If no data is available for reading, the method returns Ok(Poll::Pending) and arranges for the current task to receive a notification when the socket becomes readable or is closed.

Panics

This function will panic if called from outside of a task context.

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O on the specified portions to return immediately with an appropriate value (see the documentation of Shutdown).

pub fn nodelay(&self) -> Result<bool>

Gets the value of the TCP_NODELAY option on this socket.

For more information about this option, see set_nodelay.

pub fn set_nodelay(&self, nodelay: bool) -> Result<()>

Sets the value of the TCP_NODELAY option on this socket.

If set, this option disables the Nagle algorithm. This means that segments are always sent as soon as possible, even if there is only a small amount of data. When not set, data is buffered until there is a sufficient amount to send out, thereby avoiding the frequent sending of small packets.

pub fn recv_buffer_size(&self) -> Result<usize>

Gets the value of the SO_RCVBUF option on this socket.

For more information about this option, see set_recv_buffer_size.

pub fn set_recv_buffer_size(&self, size: usize) -> Result<()>

Sets the value of the SO_RCVBUF option on this socket.

Changes the size of the operating system's receive buffer associated with the socket.

pub fn send_buffer_size(&self) -> Result<usize>

Gets the value of the SO_SNDBUF option on this socket.

For more information about this option, see set_send_buffer.

pub fn set_send_buffer_size(&self, size: usize) -> Result<()>

Sets the value of the SO_SNDBUF option on this socket.

Changes the size of the operating system's send buffer associated with the socket.

pub fn keepalive(&self) -> Result<Option<Duration>>

Returns whether keepalive messages are enabled on this socket, and if so the duration of time between them.

For more information about this option, see set_keepalive.

pub fn set_keepalive(&self, keepalive: Option<Duration>) -> Result<()>

Sets whether keepalive messages are enabled to be sent on this socket.

On Unix, this option will set the SO_KEEPALIVE as well as the TCP_KEEPALIVE or TCP_KEEPIDLE option (depending on your platform). On Windows, this will set the SIO_KEEPALIVE_VALS option.

If None is specified then keepalive messages are disabled, otherwise the duration specified will be the time to remain idle before sending a TCP keepalive probe.

Some platforms specify this value in seconds, so sub-second specifications may be omitted.

pub fn ttl(&self) -> Result<u32>

Gets the value of the IP_TTL option for this socket.

For more information about this option, see set_ttl.

pub fn set_ttl(&self, ttl: u32) -> Result<()>

Sets the value for the IP_TTL option on this socket.

This value sets the time-to-live field that is used in every packet sent from this socket.

pub fn linger(&self) -> Result<Option<Duration>>

Reads the linger duration for this socket by getting the SO_LINGER option.

For more information about this option, see set_linger.

pub fn set_linger(&self, dur: Option<Duration>) -> Result<()>

Sets the linger duration of this socket by setting the SO_LINGER option.

This option controls the action taken when a stream has unsent messages and the stream is closed. If SO_LINGER is set, the system shall block the process until it can transmit the data or until the time expires.

If SO_LINGER is not specified, and the stream is closed, the system handles the call in a way that allows the process to continue as quickly as possible.

pub fn try_clone(&self) -> Result<TcpStream>

Creates a new independently owned handle to the underlying socket.

The returned TcpStream is a reference to the same stream that this object references. Both handles will read and write the same stream of data, and options set on one stream will be propagated to the other stream.