Struct romio::raw::PollEvented

pub struct PollEvented<E: Evented> { /* fields omitted */ }

Associates an I/O resource that implements the std::io::Read and/or std::io::Write traits with the reactor that drives it.

PollEvented uses Registration internally to take a type that implements mio::Evented as well as std::io::Read and or std::io::Write and associate it with a reactor that will drive it.

Once the mio::Evented type is wrapped by PollEvented, it can be used from within the future's execution model. As such, the PollEvented type provides AsyncRead and AsyncWrite implementations using the underlying I/O resource as well as readiness events provided by the reactor.

Note: While PollEvented is Sync (if the underlying I/O type is Sync), the caller must ensure that there are at most two tasks that use a PollEvented instance concurrently. One for reading and one for writing. While violating this requirement is "safe" from a Rust memory model point of view, it will result in unexpected behavior in the form of lost notifications and tasks hanging.

Readiness events

Besides just providing AsyncRead and AsyncWrite implementations, this type also supports access to the underlying readiness event stream. While similar in function to what Registration provides, the semantics are a bit different.

Two functions are provided to access the readiness events: poll_read_ready and poll_write_ready. These functions return the current readiness state of the PollEvented instance. If poll_read_ready indicates read readiness, immediately calling poll_read_ready again will also indicate read readiness.

When the operation is attempted and is unable to succeed due to the I/O resource not being ready, the caller must call clear_read_ready or clear_write_ready. This clears the readiness state until a new readiness event is received.

This allows the caller to implement additional functions. For example, TcpListener implements poll_accept by using poll_read_ready and clear_read_ready.

pub fn poll_accept(&mut self) -> Poll<(net::TcpStream, SocketAddr), io::Error> {
    let ready = Ready::readable();

    try_ready!(self.poll_evented.poll_read_ready(ready));

    match self.poll_evented.get_ref().accept_std() {
        Ok(pair) => Ok(Async::Ready(pair)),
        Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => {
            self.poll_evented.clear_read_ready(ready);
            Ok(Async::NotReady)
        }
        Err(e) => Err(e),
    }
}

Platform-specific events

PollEvented also allows receiving platform-specific mio::Ready events. These events are included as part of the read readiness event stream. The write readiness event stream is only for Ready::writable() events.

Methods

impl<E> PollEvented<E> where
E: Evented,

pub fn new(io: E) -> PollEvented<E>

Creates a new PollEvented associated with the default reactor.

pub fn get_ref(&self) -> &E

Returns a shared reference to the underlying I/O object this readiness stream is wrapping.

pub fn get_mut(&mut self) -> &mut E

Returns a mutable reference to the underlying I/O object this readiness stream is wrapping.

pub fn poll_read_ready(
self: Pin<&mut Self>,
cx: &mut Context
) -> Poll<Result<Ready>>

Check the I/O resource's read readiness state.

The mask argument allows specifying what readiness to notify on. This can be any value, including platform specific readiness, except writable. HUP is always implicitly included on platforms that support it.

If the resource is not ready for a read then Async::NotReady is returned and the current task is notified once a new event is received.

The I/O resource will remain in a read-ready state until readiness is cleared by calling clear_read_ready.

pub fn clear_read_ready(self: Pin<&mut Self>, cx: &mut Context) -> Result<()>

Clears the I/O resource's read readiness state and registers the current task to be notified once a read readiness event is received.

After calling this function, poll_read_ready will return NotReady until a new read readiness event has been received.

The mask argument specifies the readiness bits to clear. This may not include writable or hup.

pub fn poll_write_ready(&self, cx: &mut Context) -> Poll<Result<Ready, Error>>

Check the I/O resource's write readiness state.

This always checks for writable readiness and also checks for HUP readiness on platforms that support it.

If the resource is not ready for a write then Async::NotReady is returned and the current task is notified once a new event is received.

The I/O resource will remain in a write-ready state until readiness is cleared by calling clear_write_ready.

Panics

This function panics if:

ready contains bits besides writable and hup.

called from outside of a task context.

pub fn clear_write_ready(self: Pin<&mut Self>, cx: &mut Context) -> Result<()>

Resets the I/O resource's write readiness state and registers the current task to be notified once a write readiness event is received.

This only clears writable readiness. HUP (on platforms that support HUP) cannot be cleared as it is a final state.

After calling this function, poll_write_ready(Ready::writable()) will return NotReady until a new write readiness event has been received.

Panics

This function will panic if called from outside of a task context.