Trait romio::raw::TakeError

pub trait TakeError {
type Ok: Error + Send + Sync;
type Err: Error + Send + Sync;
    fn take_error(&self) -> Result<Option<Self::Ok>, Self::Err>;
}

Extract an error from the underlying struct that isn't propagated through regular channels.

This is common in TcpListener / UdsStream structs where this trait can be used to access the SO_ERROR option on the socket.

Both Ok and Err are error types. If no error exists take_error should return Ok(None).

Associated Types

type Ok: Error + Send + Sync

The type of successful values yielded by this trait.

type Err: Error + Send + Sync

The type of failures yielded by this trait.

Loading content...
Required methods

fn take_error(&self) -> Result<Option<Self::Ok>, Self::Err>

Return an underlying error value of the struct.

Loading content...