Struct romio::tcp::TcpStream

pub struct TcpStream { /* fields omitted */ }

A TCP stream between a local and a remote socket.

A TcpStream can either be created by connecting to an endpoint, via the connect method, or by accepting a connection from a listener. It can be read or written to using the AsyncRead, AsyncWrite, and related extension traits in futures::io.

The connection will be closed when the value is dropped. The reading and writing portions of the connection can also be shut down individually with the [shutdown] method.

Methods

impl TcpStream

Important traits for ConnectFuture
impl Future for ConnectFuture type Output = Result<TcpStream>;

pub fn connect(addr: &SocketAddr) -> ConnectFuture

Create a new TCP stream connected to the specified address.

This function will create a new TCP socket and attempt to connect it to the addr provided. The returned future will be resolved once the stream has successfully connected, or it will return an error if one occurs.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1".parse().unwrap();
await!(TcpStream::connect(&addr))

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this stream is bound to.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;
use std::net::{IpAddr, Ipv4Addr};

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

let expected = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
assert_eq!(stream.local_addr()?.ip(), expected);

pub fn peer_addr(&self) -> Result<SocketAddr>

Returns the remote address that this stream is connected to.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;
use std::net::{Ipv4Addr, SocketAddr, SocketAddrV4};

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

let expected = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8080);
assert_eq!(stream.peer_addr()?, SocketAddr::V4(expected));

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O on the specified portions to return immediately with an appropriate value (see the documentation of Shutdown).

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;
use std::net::Shutdown;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.shutdown(Shutdown::Both)?;

pub fn nodelay(&self) -> Result<bool>

Gets the value of the TCP_NODELAY option on this socket.

For more information about this option, see set_nodelay.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_nodelay(true)?;
assert_eq!(stream.nodelay()?, true);

pub fn set_nodelay(&self, nodelay: bool) -> Result<()>

Sets the value of the TCP_NODELAY option on this socket.

If set, this option disables the Nagle algorithm. This means that segments are always sent as soon as possible, even if there is only a small amount of data. When not set, data is buffered until there is a sufficient amount to send out, thereby avoiding the frequent sending of small packets.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_nodelay(true)?;

pub fn recv_buffer_size(&self) -> Result<usize>

Gets the value of the SO_RCVBUF option on this socket.

For more information about this option, see set_recv_buffer_size.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_recv_buffer_size(100);
assert_eq!(stream.recv_buffer_size()?, 100);

pub fn set_recv_buffer_size(&self, size: usize) -> Result<()>

Sets the value of the SO_RCVBUF option on this socket.

Changes the size of the operating system's receive buffer associated with the socket.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_recv_buffer_size(100);

pub fn send_buffer_size(&self) -> Result<usize>

Gets the value of the SO_SNDBUF option on this socket.

For more information about this option, see set_send_buffer.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_send_buffer_size(100);
assert_eq!(stream.send_buffer_size()?, 100);

pub fn set_send_buffer_size(&self, size: usize) -> Result<()>

Sets the value of the SO_SNDBUF option on this socket.

Changes the size of the operating system's send buffer associated with the socket.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_send_buffer_size(100);

pub fn keepalive(&self) -> Result<Option<Duration>>

Returns whether keepalive messages are enabled on this socket, and if so the duration of time between them.

For more information about this option, see set_keepalive.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;
use std::time::Duration;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_keepalive(Some(Duration::from_secs(60)))?;
assert_eq!(stream.keepalive()?, Some(Duration::from_secs(60)));

pub fn set_keepalive(&self, keepalive: Option<Duration>) -> Result<()>

Sets whether keepalive messages are enabled to be sent on this socket.

On Unix, this option will set the SO_KEEPALIVE as well as the TCP_KEEPALIVE or TCP_KEEPIDLE option (depending on your platform). On Windows, this will set the SIO_KEEPALIVE_VALS option.

If None is specified then keepalive messages are disabled, otherwise the duration specified will be the time to remain idle before sending a TCP keepalive probe.

Some platforms specify this value in seconds, so sub-second specifications may be omitted.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;
use std::time::Duration;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_keepalive(Some(Duration::from_secs(60)))?;

pub fn ttl(&self) -> Result<u32>

Gets the value of the IP_TTL option for this socket.

For more information about this option, see set_ttl.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_ttl(100)?;
assert_eq!(stream.ttl()?, 100);

pub fn set_ttl(&self, ttl: u32) -> Result<()>

Sets the value for the IP_TTL option on this socket.

This value sets the time-to-live field that is used in every packet sent from this socket.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_ttl(100)?;

pub fn linger(&self) -> Result<Option<Duration>>

Reads the linger duration for this socket by getting the SO_LINGER option.

For more information about this option, see set_linger.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;
use std::time::Duration;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_linger(Some(Duration::from_millis(100)))?;
assert_eq!(stream.linger()?, Some(Duration::from_millis(100)));

pub fn set_linger(&self, dur: Option<Duration>) -> Result<()>

Sets the linger duration of this socket by setting the SO_LINGER option.

This option controls the action taken when a stream has unsent messages and the stream is closed. If SO_LINGER is set, the system shall block the process until it can transmit the data or until the time expires.

If SO_LINGER is not specified, and the stream is closed, the system handles the call in a way that allows the process to continue as quickly as possible.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::tcp::TcpStream;
use std::time::Duration;

let addr = "127.0.0.1:8080".parse()?;
let stream = await!(TcpStream::connect(&addr))?;

stream.set_linger(Some(Duration::from_millis(100)))?;