Struct romio::uds::UnixDatagram

pub struct UnixDatagram { /* fields omitted */ }

An I/O object representing a Unix datagram socket.

Methods

impl UnixDatagram

pub fn bind(path: impl AsRef<Path>) -> Result<UnixDatagram>

Creates a new UnixDatagram bound to the specified path.

Examples

use romio::uds::UnixDatagram;

let sock = UnixDatagram::bind("/tmp/sock")?;

pub fn pair() -> Result<(UnixDatagram, UnixDatagram)>

Creates an unnamed pair of connected sockets.

This function will create a pair of interconnected Unix sockets for communicating back and forth between one another. Each socket will be associated with the event loop whose handle is also provided.

Examples

use romio::uds::UnixDatagram;

let (sock1, sock2) = UnixDatagram::pair()?;

pub fn unbound() -> Result<UnixDatagram>

Creates a new UnixDatagram which is not bound to any address.

Examples

use romio::uds::UnixDatagram;

let sock = UnixDatagram::unbound()?;

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this socket is bound to.

Examples

use romio::uds::UnixDatagram;

let stream = UnixDatagram::bind("/tmp/sock")?;
let addr = stream.local_addr()?;

pub fn peer_addr(&self) -> Result<SocketAddr>

Returns the address of this socket's peer.

The connect method will connect the socket to a peer.

Examples

use romio::uds::UnixDatagram;

let stream = UnixDatagram::bind("/tmp/sock")?;
let addr = stream.peer_addr()?;

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shut down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).

Examples

use romio::uds::UnixDatagram;
use std::net::Shutdown;

let stream = UnixDatagram::bind("/tmp/sock")?;
stream.shutdown(Shutdown::Both)?;