Struct romio::uds::UnixListener

pub struct UnixListener { /* fields omitted */ }

A Unix socket which can accept connections from other Unix sockets.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::uds::{UnixListener, UnixStream};
use futures::prelude::*;

async fn say_hello(mut stream: UnixStream) {
    await!(stream.write_all(b"Shall I hear more, or shall I speak at this?!"));
}

async fn listen() -> Result<(), Box<dyn std::error::Error + 'static>> {
    let listener = UnixListener::bind("/tmp/sock")?;
    let mut incoming = listener.incoming();

    // accept connections and process them serially
    while let Some(stream) = await!(incoming.next()) {
        await!(say_hello(stream?));
    }
    Ok(())
}

Methods

impl UnixListener

pub fn bind(
path: impl AsRef<Path>
) -> Result<UnixListener>

Creates a new UnixListener bound to the specified path.

Examples

Create a Unix Domain Socket on /tmp/sock.

use romio::uds::UnixListener;

let socket = UnixListener::bind("/tmp/sock")?;

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local socket address of this listener.

Examples

use romio::uds::UnixListener;

let socket = UnixListener::bind("/tmp/sock")?;
let addr = socket.local_addr()?;

pub fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

Examples

use romio::uds::UnixListener;

let listener = UnixListener::bind("/tmp/sock")?;
if let Ok(Some(err)) = listener.take_error() {
    println!("Got error: {:?}", err);
}

pub fn incoming(self) -> Incoming

Consumes this listener, returning a stream of the sockets this listener accepts.

This method returns an implementation of the Stream trait which resolves to the sockets the are accepted on this listener.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::uds::UnixListener;
use futures::prelude::*;

let listener = UnixListener::bind("/tmp/sock")?;
let mut incoming = listener.incoming();

// accept connections and process them serially
while let Some(stream) = await!(incoming.next()) {
    match stream {
        Ok(stream) => {
            println!("new client!");
        },
        Err(e) => { /* connection failed */ }
    }
}