Struct romio::uds::UnixStream

pub struct UnixStream { /* fields omitted */ }

A structure representing a connected Unix socket.

This socket can be connected directly with UnixStream::connect or accepted from a listener with UnixListener::incoming. Additionally, a pair of anonymous Unix sockets can be created with UnixStream::pair.

Methods

impl UnixStream

pub fn connect(
path: impl AsRef<Path>
) -> ConnectFuture

Connects to the socket named by path.

This function will create a new Unix socket and connect to the path specified, associating the returned stream with the default event loop's handle.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::uds::UnixStream;

let stream = await!(UnixStream::connect("/tmp/sock"));

pub fn pair() -> Result<(UnixStream, UnixStream)>

Creates an unnamed pair of connected sockets.

This function will create a pair of interconnected Unix sockets for communicating back and forth between one another. Each socket will be associated with the event loop whose handle is also provided.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::uds::UnixStream;

let (sock1, sock2) = UnixStream::pair()?;

pub fn poll_read_ready(&self, lw: &LocalWaker) -> Poll<Result<Ready>>

Test whether this socket is ready to be read or not.

pub fn poll_write_ready(&self, lw: &LocalWaker) -> Poll<Result<Ready>>

Test whether this socket is ready to be written to or not.

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the socket address of the local half of this connection.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::uds::UnixStream;

let stream = await!(UnixStream::connect("/tmp/sock"))?;
let addr = stream.local_addr()?;

pub fn peer_addr(&self) -> Result<SocketAddr>

Returns the socket address of the remote half of this connection.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::uds::UnixStream;

let stream = await!(UnixStream::connect("/tmp/sock"))?;
let addr = stream.peer_addr()?;

pub fn peer_cred(&self) -> Result<UCred>

Returns effective credentials of the process which called connect or socketpair.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::uds::UnixStream;

let stream = await!(UnixStream::connect("/tmp/sock"))?;
let cred = stream.peer_cred()?;

pub fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

Examples

#![feature(async_await, await_macro, futures_api)]
use romio::uds::UnixStream;

let stream = await!(UnixStream::connect("/tmp/sock"))?;
if let Ok(Some(err)) = stream.take_error() {
    println!("Got error: {:?}", err);
}

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).

#![feature(async_await, await_macro, futures_api)]
use romio::uds::UnixStream;
use std::net::Shutdown;

let stream = await!(UnixStream::connect("/tmp/sock"))?;
stream.shutdown(Shutdown::Both)?;